Additional Organizer Code Validation in Service
I added an if statement to the get event service that checks to make sure the organizer code in the request is the same as the one listed for the event in the database. This will prevent a user from posting the edit event function directly without the proper organizer code.